Splitting popup method in his own source.
